{{cookiecutter.repo_name}}/requirements.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

boto==2.38.0
dj-database-url==0.3.0
django-storages-redux==1.2.3
django-debug-toolbar==1.3.0
django-cms==3.1
django-extensions==1.5.1
django-model-utils==2.2
djangocms-column==1.5
djangocms-forms==0.1.4
djangocms-link==1.6.1
djangocms-picture==0.1
djangocms-text-ckeditor==2.5.0
djrill==1.4.0
gunicorn==19.3.0
Pillow==2.8.2
psycopg2==2.6.1
sorl-thumbnail==12.2
whitenoise==1.0.6
